Move Rule helpers to services.
